+++ frontend/src/embedding-sdk-bundle/lib/sdk-question/run-question-query.ts
@@ -1,7 +1,7 @@
 import { getGuestEmbedFilteredParameters } from "embedding-sdk-bundle/lib/get-guest-embed-filtered-parameters";
 import type { SdkQuestionState } from "embedding-sdk-bundle/types/question";
+import { runQuestionQuery } from "metabase/querying/run-query";
 import type { Dispatch } from "metabase/redux/store";
-import { runQuestionQuery } from "metabase/services";
 import { getSensibleDisplays } from "metabase/visualizations";
 import type Question from "metabase-lib/v1/Question";
 import type { ParameterValuesMap } from "metabase-types/api";
